// src/app/services/openapi-client/model/updateOrderStatusDto.ts
/**
 * PaintDepotApi
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface UpdateOrderStatusDto { 
    status: string;
    trackingNumber?: string | null;
    adminNotes?: string | null;
}